# specs/000-p0-scaffold/contracts/proxy.openapi.yaml
openapi: 3.0.0
info:
  title: bus_dashboard BFF proxy surface (P0)
  version: 0.0.0-p0
  description: >-
    Same-origin proxy. Browser NEVER calls BUS_API_URL directly. JWTs in
    httpOnly cookies. Envelope {statusCode,data} unwrapped once server-side.
    Cursor pages {items,nextCursor} passed through untouched. Backend source:
    ../bus_api/docs/openapi.json (do not modify backend).
servers:
  - url: /api
    description: Next.js BFF (same origin)
paths:
...
  /auth/login:
    post:
      operationId: proxyLogin
      summary: Proxied POST /auth/login + set httpOnly cookies
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                rememberMe: { type: boolean }
...